[clean] コードを整理する(リネーム+フォーマット)
